+++ .eslintrc.js
@@ -1,0 +1,24 @@
+module.exports = {
+  'env': {
+    'browser': true,
+    'es6': true,
+  },
+  'extends': [
+    'plugin:vue/essential',
+    'google',
+  ],
+  'globals': {
+    'Atomics': 'readonly',
+    'SharedArrayBuffer': 'readonly',
+  },
+  'parserOptions': {
+    'ecmaVersion': 2018,
+    'sourceType': 'module',
+  },
+  'plugins': [
+    'vue',
+  ],
+  'rules': {
+    'require-jsdoc': 0,
+  },
+};
