+++ public/index.html
@@ -69,7 +69,7 @@
   <div id="footer" class="ui">
     <ul>
       <li>
-        &copy;2020 <a href="https://randychen.tk" target="_blank" rel="noopener">SuperSonic</a>.
+        &copy;2021 <a href="https://randychen.tk" target="_blank" rel="noopener">SuperSonic</a>.
       </li>
     </ul>
   </div>
